(agentxmpp) restart the configurator if the file cluster.ini is missing
When an agent starts it tries to connect to the relayserver agent. If
this one does not respond, it looks in the cluster.ini file for an
alternative.

Until now, if we didn't have a cluster.in file, it was trying to connect
to the same relayserver in an infinite loop.

Now, if there is no cluster.ini file, it restarts the configurator and
get a new relayserver from the master.