(Agents update) Change how we handle agent updates. Before the agents were updated all along, without limits of numbers.
Now we have an `update` table in the database.
We add inside it the machines who need to be updated and a plugin will
monitor this and allow machines to update depending of parameters ( like
the number of allowed updates at once ).

Functions added:
- __updatemachine
- update_update_machine
- setUpdate_machine
- getUpdate_machine

Documentation added in the following functions:
- __updatemachine
- update_update_machine
- setUpdate_machine
- getUpdate_machine