(utils) Fix finding the hosts
In some circunstances, we failed to define the dns of the main or
relayserver.

Now we try the Dns Resolution and if we fail we add it directly in the
hosts file of the client.

This way we do not rely on the DNS resolution and do not fail anymore.