(deploy) Do not stop missingagent deploiements.
When using the "Stop deploy" button, we had a backtrace when we had
machine a "missingagent" status.

Now do not stop depoloiements for "missingagent" deploiements as they
are already stopped and the jid domain  can be wrong, as it is a fake
domain.

In addition, we do not use the domain for this kind of machines in the
deploy SQL base

win-3.Unkown OS is now win-3