(CDN) Add new feature:
Add the possibility to deploy using a CDN. (#131)

The relay agent will check if cdn_enable is True. If yes, we will use
the URL + Token ( If not, we will deploy like before ).

Port the package watching for CDN and allo to synchronize the package
to the CDN on package changes.