(feature) CDN: Check for integrity
Now that we can use CDN like Akamai to download packages we now add the
possibilty to check the integrity of the downloaded packages.

We provide a hash of the packages and we compare.

If we don't have an available Hash we fail with the error : ERROR HASH MISSING
And if the Hash is not correct we fail with the error: ABORT HASH INVALID