(configurator) Make sure we have an IP for the xmpp server
In some conditions confserver was empty, and as it was on a while true
loop it was forever.

Now if confserver is empty we reread the configuration to have the
correct information.

This can happen when we want to read the configuration at the same as we
rotate the logs.