(launcher) Test if the config file is a valid template file and add new BOOL_FILE_INSTALL option.
When we reinstall the agent, we add a BOOL_FILE_INSTALL file.

If it is available we reconfigure the agent, as this is new install on
an existing one.

When we reinstall the agent, we got an issue with the rescue.
It considered that the install was incorrect because the configuration
file was a template.
Now we test if this is a template file. If this is a valid template file
we restart the configurator and if not we restore the rescue image.