(inventory plugin) Do not raise a backtrace if the inventory file does not exists.
We have cases where the inventory file does not exists.
One "normal" and one dysfunctionnal.

The normal one is when the medulla agent just started. He needs time to
asks Fusioninventory to create the inventory file so the first use of
the plugin can be too early, but this is not a bug.

But if the medulla agent is started for a while this should not happen.

To help debugging let's explain this on the warning :-)