(syncthingapirest) Do not delete pulse device in syncthing configuration
When we configure the ARS we remove the device of the server we are on.
But the problem is the main server does not use the hostname of the
server but pulse as device name instead.

Deleting the device result in a bad configuration after as the pulse
device does not exist anymore.

I keep the deletion of the device for the other servers ( but the main
server).