(updatesettings) Remove liststartplugin from default configuration
Before, we didn't had the availability to exclude plugins, so we
activated plugins one by one.

Now we have the choice to enable all or only some. If the
liststartplugin configuration option does not exist we consider that we
enable all the plugins.

This commit allow to remove the previous liststartplugin configuration
from older configuration files to enable all.

This allow to enable new plugins that weren't available before.