(python path) Install python in python3 folder instead of versionned python311
The agent installed on python folder. For the moment we only support
python 3.11 but when 3.12 will supported by us and 3.11 dead upstream we
will want to update.

With a non versioned folder this will ease the update as we will only
have to update the python + some modules python version depending.