(medullanetnotify) Fix medullanetnotify service installation
We couldn't install the service because pythonservices.exe was already
on the system and already on use by the medullaagent service.

To fix this, we need to have a copy of this file, named
medullanetnotify.exe. This way pywin32 will install this new file on the
correct location.

(cherry picked from commit 400f07f276dc96945a83983faab246d25598b0ad)