(agent machine) More robust machine agent.
In some cases, the agentconf.ini file is corrupted,
i.e. it does not contain all the required information.

The reasons for this file corruption are not yet known,
but in order to remedy the situation, we have created a
sample file which is used when this problem is detected.

In this commit, we look for the minimum information
required for the agent to function correctly.