(installer) Install a new file called agentconfig.ini.tpl
This file won't be used by medulla when running.
In a next commit we will add a change:
-> When medulla starts it will check for some sections on
agentconfig.ini.
If ok nothing will change, but if we miss some sections that can
prevent the agent to start, we will copy agentconfig.ini.tpl into agentconfig.ini

Next steps will be to do it for some important configuration files too.