(launcher) Use the template file if the config file is missing
If the agentconf file is missing, we use the template file and we recopy
it on the correct location
Important: we only use this for the machine agent.

(cherry picked from commit bde19d0a81b57a3798b8477198be7d0c376c2d4a)

Rename the isTemplateConfFile into isValidConfFile

Fix case when the conf file is not valid

Sometimes the conffile may not be valid ( empty, corrupted, etc).

In this case we reuse the template file if possible