(updatetightvnc plugin) Remove use of the deprecated distutils.version module
As the new version of tightvnc is 2.8.84.0 and as distutils.version is
deprecated we needed to get rid of it.

Now we use the new packaging module that provide the Version() function
which is recommended for this.

(cherry picked from commit d13715a645ef92164850762f2a160ae28d9d34f4)