(Extract Drivers) Launch the extract driver deployment if not done yet
On Windows OS machine, if the machine hasn't the ExtractedDrivers registry keys, means the extraction of the drivers has not been automatically done.
On Windows OS machine, if the macine has the ExtractedDrivers registry keys, means the extraction of the drivers has been automatically done.
On other OS machine, the automatic extraction drivers is not handled.